Refactor: Extract LabeledItem component to its own file
